compose: Remove remains of "scroll to message" banner.
It seems like there used to be a link in the scroll to message
banner, but there is no longer any action link on that banner.

This commit contains two changes:

(1) No longer passing the narrow url to the render function.
Since there's no link class specified, the url won't show up
at all so there's no reason to send the url through.

(2) Remove the click handler for what was presumably the link
class before it got removed.