buddy list: Rename user_presences to narrow-user-presences.
This switches to our preferred dash-separated classnames
instead of underscore-separated, and also updates the
classname to be differentiable to differentiable from
users that aren't part of the current narrow.

Until the next commit that splits the buddy list, the
name is a bit inaccurate since it still contains all
the users.