settings: Hide invisible sections on narrow screens instead of layering.
This is a much cleaner way to sometimes have the left pane be the
full width and sometimes the right pane. Previously it was set up
so that both were visible, but they were layered with z-index and
absolute positioning. This commit refactors the stream and usergroup
settings but not the main settings page yet.

This commit should have no visible changes.