left_sidebar: Remove (already hidden) #streams_header and related code.
This is an attempt to split up work into separate commits and avoid
a huge diff, but we'll see if it's actually helpful. I imagine we'll
at least want to merge it with other commits when we eventually
merge the PR.

Everything removed here that's still relevant (unread counts, special
logic for spectator view) is added back in over the next several
commits.